requirements.txt
# 企业知识库 RAG 问答机器人 - 依赖
# Python 3.10+

# 核心框架
langchain>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0

# 向量与嵌入（本地可选，推荐 langchain-huggingface 替代已弃用的 community）
chromadb>=0.5.0
sentence-transformers>=3.0.0
langchain-huggingface>=0.1.0

# 文档解析
pypdf>=4.0.0
python-docx>=1.0.0
unstructured>=0.15.0
# 注意：如需支持旧版 .doc 文件，可能需要额外安装: pip install 'unstructured[local]'
# 或更简单：将 .doc 文件转换为 .docx 格式

# LLM 接入（按需选用，OpenAI 兼容 API）
openai>=1.0.0
langchain-openai>=0.2.0
httpx>=0.27.0

# API 与 Web
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9

# 配置与工具
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
loguru>=0.7.0
